Perform various modification for cleaning up the GPU solvers
