+++ examples/heatConductionScript/heatConduction2DFin/HeatConduction2DFinWebgpu.html
