+++ src/methods/webgpuComputeEngineScript.js
@@ -1,3 +1,14 @@
+//   ______ ______           _____           _       _     //
+//  |  ____|  ____|   /\    / ____|         (_)     | |    //
+//  | |__  | |__     /  \  | (___   ___ ____ _ ____ | |_   //
+//  |  __| |  __|   / /\ \  \___ \ / __|  __| |  _ \| __|  //
+//  | |    | |____ / ____ \ ____) | (__| |  | | |_) | |    //
+//  |_|    |______/_/    \_\_____/ \___|_|  |_|  __/| |    //
+//                                            | |   | |    //
+//                                            |_|   | |_   //
+//       Website: https://feascript.com/             \__|  //
+
+// External imports
 import * as ti from '../vendor/taichi.esm.js';
 
 export class WebGPUComputeEngine {
@@ -6,7 +17,7 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Initialize Taichi
+   * Function to initialize the WebGPU compute engine
    */
   async initialize() {
     if (this.initialized) return;
@@ -16,14 +27,22 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Read data from a buffer (compatibility method)
+   * Function to read data from a buffer
+   * @param {object} data - The buffer to read from
+   * @param {number} size - The size of the data to read
+   * @param {object} [type=Float32Array] - The type of the data array
+   * @returns {object} The data read from the buffer
    */
   async readBuffer(data, size, type = Float32Array) {
     return data;
   }
 
   /**
-   * Matrix-vector multiplication: y = A * x
+   * Function to perform matrix-vector multiplication: y = A * x
+   * @param {array} A - The matrix
+   * @param {array} x - The vector
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async matVecMul(A, x, resultBuffer = null) {
     const n = A.length;
@@ -37,7 +56,7 @@ export class WebGPUComputeEngine {
     matrixField.fromArray(flatMatrix);
     vectorField.fromArray(x);
 
-    ti.addToKernelScope({matrixField, vectorField, resultField});
+    ti.addToKernelScope({ matrixField, vectorField, resultField });
 
     ti.kernel((n, m) => {
       for (let i of ti.ndrange(n)) {
@@ -54,7 +73,11 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Vector addition: result = a + b
+   * Function to perform vector addition: result = a + b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async vecAdd(a, b, resultBuffer = null) {
     const n = a.length;
@@ -65,7 +88,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     bField.fromArray(b);
 
-    ti.addToKernelScope({aField, bField, resultField});
+    ti.addToKernelScope({ aField, bField, resultField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -77,8 +100,12 @@ export class WebGPUComputeEngine {
     return result;
   }
 
-    /**
-   * Vector subtraction: result = a - b
+  /**
+   * Function to perform vector subtraction: result = a - b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async vecSub(a, b, resultBuffer = null) {
     const n = a.length;
@@ -89,7 +116,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     bField.fromArray(b);
 
-    ti.addToKernelScope({aField, bField, resultField});
+    ti.addToKernelScope({ aField, bField, resultField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -102,7 +129,11 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Element-wise vector multiplication: result = a * b
+   * Function to perform element-wise vector multiplication: result = a * b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async vecMul(a, b, resultBuffer = null) {
     const n = a.length;
@@ -113,7 +144,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     bField.fromArray(b);
 
-    ti.addToKernelScope({aField, bField, resultField});
+    ti.addToKernelScope({ aField, bField, resultField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -126,7 +157,11 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Vector division: result = a / b
+   * Function to perform vector division: result = a / b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async vecDiv(a, b, resultBuffer = null) {
     const n = a.length;
@@ -137,7 +172,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     bField.fromArray(b);
 
-    ti.addToKernelScope({aField, bField, resultField});
+    ti.addToKernelScope({ aField, bField, resultField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -150,7 +185,9 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Find maximum value in vector
+   * Function to find the maximum value in a vector
+   * @param {array} vector - The input vector
+   * @returns {number} The maximum value in the vector
    */
   async vecMax(vector) {
     const n = vector.length;
@@ -160,7 +197,7 @@ export class WebGPUComputeEngine {
     vectorField.fromArray(vector);
     maxField.fromArray([-1e10]);
 
-    ti.addToKernelScope({vectorField, maxField});
+    ti.addToKernelScope({ vectorField, maxField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -174,8 +211,14 @@ export class WebGPUComputeEngine {
     return Math.max(...result);
   }
 
-    /**
-   * Jacobi update: x_new = D^(-1) * (b - R * x_old)
+  /**
+   * Function to perform a Jacobi update step: x_new = D^(-1) * (b - R * x_old)
+   * @param {array} A - The system matrix
+   * @param {array} b - The right-hand side vector
+   * @param {array} x - The current solution vector
+   * @param {number} [omega=1.0] - The relaxation factor
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The updated solution vector
    */
   async jacobiUpdate(A, b, x, omega = 1.0, resultBuffer = null) {
     const n = x.length;
@@ -189,7 +232,7 @@ export class WebGPUComputeEngine {
     bField.fromArray(b);
     xField.fromArray(x);
 
-    ti.addToKernelScope({AField, bField, xField, resultField});
+    ti.addToKernelScope({ AField, bField, xField, resultField });
 
     ti.kernel((n, omega) => {
       for (let i = 0; i < n; i++) {
@@ -208,7 +251,11 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Element-wise operation on vector: result[i] = op(a[i])
+   * Function to perform an element-wise operation on a vector: result[i] = op(a[i])
+   * @param {array} a - The input vector
+   * @param {string} op - The operation to perform ('abs', 'sqrt', 'exp', 'log', 'sin', 'cos', 'tan')
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async elementWiseOp(a, op, resultBuffer = null) {
     const n = a.length;
@@ -217,7 +264,7 @@ export class WebGPUComputeEngine {
 
     aField.fromArray(a);
 
-    ti.addToKernelScope({aField, resultField});
+    ti.addToKernelScope({ aField, resultField });
 
     ti.kernel((n, op) => {
       for (let i of ti.ndrange(n)) {
@@ -246,7 +293,12 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Scalar operation on vector: result[i] = op(a[i], scalar)
+   * Function to perform a scalar operation on a vector: result[i] = op(a[i], scalar)
+   * @param {array} a - The input vector
+   * @param {number} scalar - The scalar value
+   * @param {string} op - The operation to perform ('add', 'mul', 'div', 'pow')
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async scalarOp(a, scalar, op, resultBuffer = null) {
     const n = a.length;
@@ -255,7 +307,7 @@ export class WebGPUComputeEngine {
 
     aField.fromArray(a);
 
-    ti.addToKernelScope({aField, resultField});
+    ti.addToKernelScope({ aField, resultField });
 
     ti.kernel((n, scalar, op) => {
       for (let i of ti.ndrange(n)) {
@@ -278,11 +330,10 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Vector L2 norm: result = ||a||_2
-   */
-
-  /**
-   * Vector addition: c = a + b
+   * Function to perform vector addition: c = a + b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @returns {array} The resulting vector
    */
   async vecAdd(a, b) {
     const n = a.length;
@@ -293,7 +344,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     bField.fromArray(b);
 
-    ti.addToKernelScope({aField, bField, cField});
+    ti.addToKernelScope({ aField, bField, cField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -306,7 +357,10 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Element-wise vector multiplication: c = a .* b
+   * Function to perform element-wise vector multiplication: c = a .* b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @returns {array} The resulting vector
    */
   async vecMul(a, b) {
     const n = a.length;
@@ -317,7 +371,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     bField.fromArray(b);
 
-    ti.addToKernelScope({aField, bField, cField});
+    ti.addToKernelScope({ aField, bField, cField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -329,7 +383,10 @@ export class WebGPUComputeEngine {
     return result;
   }
   /**
-   * Vector dot product: result = a · b
+   * Function to compute the dot product of two vectors: result = a · b
+   * @param {array} a - The first vector
+   * @param {array} b - The second vector
+   * @returns {number} The dot product
    */
   async dotProduct(a, b) {
     const n = a.length;
@@ -341,7 +398,7 @@ export class WebGPUComputeEngine {
     bField.fromArray(b);
     resultField.fromArray([0]);
 
-    ti.addToKernelScope({aField, bField, resultField});
+    ti.addToKernelScope({ aField, bField, resultField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -354,7 +411,9 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Vector L2 norm: result = ||a||_2
+   * Function to compute the L2 norm of a vector: result = ||a||_2
+   * @param {array} a - The input vector
+   * @returns {number} The L2 norm of the vector
    */
   async norm(a) {
     const n = a.length;
@@ -364,7 +423,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     resultField.fromArray([0]);
 
-    ti.addToKernelScope({aField, resultField});
+    ti.addToKernelScope({ aField, resultField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -380,6 +439,11 @@ export class WebGPUComputeEngine {
     return result[0];
   }
 
+  /**
+   * Function to normalize a vector
+   * @param {array} a - The vector to normalize
+   * @returns {array} The normalized vector
+   */
   async normalize(a) {
     const n = a.length;
     const aField = ti.field(ti.f32, [n]);
@@ -389,7 +453,7 @@ export class WebGPUComputeEngine {
     aField.fromArray(a);
     tempField.fromArray([0]);
 
-    ti.addToKernelScope({aField, resultField, tempField});
+    ti.addToKernelScope({ aField, resultField, tempField });
 
     ti.kernel((n) => {
       for (let i of ti.ndrange(n)) {
@@ -409,41 +473,51 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Matrix-matrix multiplication: C = A * B
+   * Function to perform matrix-matrix multiplication: C = A * B
+   * @param {array} A - The first matrix
+   * @param {array} B - The second matrix
+   * @returns {array} The resulting matrix
    */
-  /**
-   * Matrix-vector multiplication: y = A * x
-   */
-  async matVecMul(A, x) {
+  async matMatMul(A, B) {
     const m = A.length;
-    const n = A[0].length;
+    const k = A[0].length;
+    const n = B[0].length;
 
-    const AField = ti.field(ti.f32, [m * n]);
-    const xField = ti.field(ti.f32, [n]);
-    const yField = ti.field(ti.f32, [m]);
+    const AField = ti.field(ti.f32, [m * k]);
+    const BField = ti.field(ti.f32, [k * n]);
+    const CField = ti.field(ti.f32, [m * n]);
 
-    const flatA = A.flat();
-    AField.fromArray(flatA);
-    xField.fromArray(x);
+    AField.fromArray(A.flat());
+    BField.fromArray(B.flat());
 
-    ti.addToKernelScope({AField, xField, yField});
+    ti.addToKernelScope({ AField, BField, CField });
 
-    ti.kernel((m, n) => {
+    ti.kernel((m, n, k) => {
       for (let i of ti.ndrange(m)) {
-        let sum = 0.0;
         for (let j of ti.ndrange(n)) {
-          sum += AField[ti.i32(i) * ti.i32(n) + ti.i32(j)] * xField[j];
+          let sum = 0.0;
+          for (let p of ti.ndrange(k)) {
+            sum += AField[ti.i32(i) * ti.i32(k) + ti.i32(p)] *
+              BField[ti.i32(p) * ti.i32(n) + ti.i32(j)];
+          }
+          CField[ti.i32(i) * ti.i32(n) + ti.i32(j)] = sum;
         }
-        yField[i] = sum;
       }
-    })(m, n);
+    })(m, n, k);
 
-    const result = await yField.toArray();
-    return result;
+    const flatC = await CField.toArray();
+    const C = [];
+    for (let i = 0; i < m; i++) {
+      C.push(flatC.slice(i * n, (i + 1) * n));
+    }
+    return C;
   }
 
   /**
-   * Matrix transpose: B = A^T
+   * Function to transpose a matrix: B = A^T
+   * @param {array} A - The matrix to transpose
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The transposed matrix
    */
   async transpose(A, resultBuffer = null) {
     const m = A.length;
@@ -455,7 +529,7 @@ export class WebGPUComputeEngine {
     const flatA = A.flat();
     AField.fromArray(flatA);
 
-    ti.addToKernelScope({AField, BField});
+    ti.addToKernelScope({ AField, BField });
 
     ti.kernel((m, n) => {
       for (let i of ti.ndrange(m)) {
@@ -475,7 +549,10 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Extract diagonal elements: result = diag(A)
+   * Function to extract the diagonal elements of a matrix: result = diag(A)
+   * @param {array} A - The input matrix
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The vector of diagonal elements
    */
   async diagonal(A, resultBuffer = null) {
     const n = Math.min(A.length, A[0].length);
@@ -486,7 +563,7 @@ export class WebGPUComputeEngine {
     const flatA = A.flat();
     AField.fromArray(flatA);
 
-    ti.addToKernelScope({AField, resultField});
+    ti.addToKernelScope({ AField, resultField });
 
     ti.kernel((n, cols) => {
       for (let i of ti.ndrange(n)) {
@@ -499,8 +576,11 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Sparse matrix-vector multiplication using CSR format
-   * sparseMatrix should have: {values, col_indices, row_indices, rows, cols}
+   * Function to perform sparse matrix-vector multiplication using CSR format
+   * @param {object} sparseMatrix - Object with {values, col_indices, row_indices, rows, cols}
+   * @param {array} x - The vector to multiply with
+   * @param {object} [resultBuffer=null] - Optional buffer to store the result
+   * @returns {array} The resulting vector
    */
   async sparseMatVecMul(sparseMatrix, x, resultBuffer = null) {
     const { values, col_indices, row_indices, rows, cols } = sparseMatrix;
@@ -518,7 +598,7 @@ export class WebGPUComputeEngine {
     xField.fromArray(x);
     resultField.fromArray(new Array(rows).fill(0));
 
-    ti.addToKernelScope({valuesField, colIndicesField, rowIndicesField, xField, resultField});
+    ti.addToKernelScope({ valuesField, colIndicesField, rowIndicesField, xField, resultField });
 
     ti.kernel((nnz) => {
       for (let k of ti.ndrange(nnz)) {
@@ -531,7 +611,9 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Deep copy vector or matrix
+   * Function to create a deep copy of a vector or matrix
+   * @param {array} data - The vector or matrix to copy
+   * @returns {array} A deep copy of the input data
    */
   async copy(data) {
     if (Array.isArray(data[0])) {
@@ -544,7 +626,7 @@ export class WebGPUComputeEngine {
       const flatA = data.flat();
       AField.fromArray(flatA);
 
-      ti.addToKernelScope({AField, resultField});
+      ti.addToKernelScope({ AField, resultField });
 
       ti.kernel((total) => {
         for (let i of ti.ndrange(total)) {
@@ -567,7 +649,7 @@ export class WebGPUComputeEngine {
 
       aField.fromArray(data);
 
-      ti.addToKernelScope({aField, resultField});
+      ti.addToKernelScope({ aField, resultField });
 
       ti.kernel((n) => {
         for (let i of ti.ndrange(n)) {
@@ -581,7 +663,10 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Fill vector or matrix with constant value
+   * Function to fill a vector or matrix with a constant value
+   * @param {array} data - The vector or matrix to fill (determines size)
+   * @param {number} value - The value to fill with
+   * @returns {array} The filled vector or matrix
    */
   async fill(data, value) {
     if (Array.isArray(data[0])) {
@@ -590,7 +675,7 @@ export class WebGPUComputeEngine {
       const n = data[0].length;
       const resultField = ti.field(ti.f32, [m * n]);
 
-      ti.addToKernelScope({resultField});
+      ti.addToKernelScope({ resultField });
 
       ti.kernel((total, value) => {
         for (let i of ti.ndrange(total)) {
@@ -610,7 +695,7 @@ export class WebGPUComputeEngine {
       const n = data.length;
       const resultField = ti.field(ti.f32, [n]);
 
-      ti.addToKernelScope({resultField});
+      ti.addToKernelScope({ resultField });
 
       ti.kernel((n, value) => {
         for (let i of ti.ndrange(n)) {
@@ -624,7 +709,10 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Scale vector or matrix by scalar
+   * Function to scale a vector or matrix by a scalar
+   * @param {array} data - The vector or matrix to scale
+   * @param {number} scalar - The scalar value to scale by
+   * @returns {array} The scaled vector or matrix
    */
   async scale(data, scalar) {
     if (Array.isArray(data[0])) {
@@ -637,7 +725,7 @@ export class WebGPUComputeEngine {
       const flatA = data.flat();
       AField.fromArray(flatA);
 
-      ti.addToKernelScope({AField, resultField});
+      ti.addToKernelScope({ AField, resultField });
 
       ti.kernel((total, scalar) => {
         for (let i of ti.ndrange(total)) {
@@ -660,7 +748,7 @@ export class WebGPUComputeEngine {
 
       aField.fromArray(data);
 
-      ti.addToKernelScope({aField, resultField});
+      ti.addToKernelScope({ aField, resultField });
 
       ti.kernel((n, scalar) => {
         for (let i of ti.ndrange(n)) {
@@ -674,7 +762,11 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Compute residual vector: r = b - A * x
+   * Function to compute the residual vector: r = b - A * x
+   * @param {array} A - The system matrix
+   * @param {array} x - The solution vector
+   * @param {array} b - The right-hand side vector
+   * @returns {array} The residual vector
    */
   async residual(A, x, b) {
     const Ax = await this.matVecMul(A, x);
@@ -682,8 +774,12 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Apply preconditioner: solve M * z = r for z
-   * type: 'jacobi' or 'ssor'
+   * Function to apply a preconditioner: solve M * z = r for z
+   * @param {array} A - The system matrix
+   * @param {array} r - The residual vector
+   * @param {string} [type='jacobi'] - The type of preconditioner ('jacobi' or 'ssor')
+   * @param {number} [omega=1.0] - The relaxation factor for SSOR
+   * @returns {array} The vector z
    */
   async preconditioner(A, r, type = 'jacobi', omega = 1.0) {
     const n = r.length;
@@ -698,7 +794,7 @@ export class WebGPUComputeEngine {
       // This is a basic implementation; full SSOR would require forward/backward sweeps
       const diag = await this.diagonal(A);
       const invDiag = diag.map(d => omega / d);
-      
+
       // For simplicity, approximate with Jacobi-like application
       // Full SSOR would need iterative application or matrix splitting
       return this.vecMul(invDiag, r);
@@ -708,10 +804,16 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Conjugate Gradient solver: solve A * x = b
-   * Returns approximate solution x
+   * Function to solve a system of linear equations using the Conjugate Gradient method (GPU asynchronous version)
+   * @param {array} A - The system matrix
+   * @param {array} b - The right-hand side vector
+   * @param {array} [x0=null] - The initial guess for the solution
+   * @param {number} [tol=1e-6] - The convergence tolerance
+   * @param {number} [maxIter=1000] - The maximum number of iterations
+   * @param {string} [preconditionerType=null] - The type of preconditioner to use
+   * @returns {array} The solution vector
    */
-  async conjugateGradient(A, b, x0 = null, tol = 1e-6, maxIter = 1000, preconditionerType = null) {
+  async webgpuConjugateGradientSolver(A, b, x0 = null, tol = 1e-6, maxIter = 1000, preconditionerType = null) {
     const n = b.length;
     let x = x0 ? await this.copy(x0) : new Array(n).fill(0.0);
 
@@ -786,10 +888,16 @@ export class WebGPUComputeEngine {
   }
 
   /**
-   * Jacobi iterative solver: solve A * x = b
-   * Returns approximate solution x
+   * Function to solve a system of linear equations using the Jacobi iterative method (GPU asynchronous version)
+   *
+   * @param {array} A - The coefficient matrix (must be square)
+   * @param {array} b - The right-hand side vector
+   * @param {array} x0 - Initial guess for solution vector
+   * @param {number} [maxIter=1000] - Maximum number of iterations
+   * @param {number} [tol=1e-6] - Convergence tolerance
+   * @returns {array} The solution vector
    */
-  async jacobiSolve(A, b, x0, maxIter = 1000, tol = 1e-6) {
+  async webgpuJacobiSolver(A, b, x0, maxIter = 1000, tol = 1e-6) {
     const n = b.length;
     let x = await this.copy(x0);
     let x_new = await this.copy(x0);
@@ -823,6 +931,9 @@ export class WebGPUComputeEngine {
     return x;
   }
 
+  /**
+   * Function to destroy the compute engine and clean up resources
+   */
   async destroy() {
     if (this.initialized) {
       // Clean up Taichi.js resources and WebGPU context
