+++ src/models/frontPropagationScript.js
